refactor: 타입 별도의 폴더로 분리
