+++ src/constants/tmdbAPI.ts
@@ -1,0 +1,3 @@
+const TMDB_API = {
+  popular: ``,
+};
