+++ src/render/setupPopularMovie.ts
@@ -1,5 +1,5 @@
 import moveList from '../components/MovieList';
-import { FetchStandard } from '../app';
+import { FetchStandard } from '../types/fetcherType';
 
 export default (fetchStandard: FetchStandard) => {
   moveList.setListName(fetchStandard.type);
